Add customizable effect for consumed item (in addition to the thirst effect) + associated tooltips + configuration files Fix morphine consumption item Modify effect tooltip for healing items Modify "already applied" tooltip & "show recipes" tooltip on sewing table screen Create new json file to group all healing parameters Enable all items to heal the body based on this json file Adapt the tooltip + api + config Add nbt specifications in thirst consumables json file Adapt tooltip & events to tackle nbt checks + fallback to hydrationPurity tag if found Rename config tickRate to tempTickTime Remove blur effect when in creative / spectator mode

+++ src/main/java/sfiomn/legendarysurvivaloverhaul/common/integration/sereneseasons/SereneSeasonsModifier.java
@@ -187,15 +187,15 @@ public class SereneSeasonsModifier extends ModifierBase
 		return applyUndergroundEffect(value, world, pos);
 	}
 
-	private float getSeasonModifier(float previousSeasonModifier, float currentSeasonModifier, float nextSeasonModifier, int time, int subSeasonDuration) {
+	private float getSeasonModifier(double previousSeasonModifier, double currentSeasonModifier, double nextSeasonModifier, int time, int subSeasonDuration) {
 		return time < subSeasonDuration / 2 ?
 				calculateSinusoidalBetweenSeasons(previousSeasonModifier, currentSeasonModifier, time + (subSeasonDuration / 2), subSeasonDuration):
 				calculateSinusoidalBetweenSeasons(currentSeasonModifier, nextSeasonModifier, time - (subSeasonDuration / 2), subSeasonDuration);
 	}
 
-	private float calculateSinusoidalBetweenSeasons(float previousSeasonModifier, float nextSeasonModifier, int time, int subSeasonDuration) {
-		float tempDiff = nextSeasonModifier - previousSeasonModifier;
-		float seasonModifier = (float) (Math.sin(((time * Math.PI) / subSeasonDuration) - (Math.PI / 2)) + 1) * (tempDiff / 2) + previousSeasonModifier;
-		return MathUtil.round(seasonModifier, 2);
+	private float calculateSinusoidalBetweenSeasons(double previousSeasonModifier, double nextSeasonModifier, int time, int subSeasonDuration) {
+		double tempDiff = nextSeasonModifier - previousSeasonModifier;
+		double seasonModifier = (float) (Math.sin(((time * Math.PI) / subSeasonDuration) - (Math.PI / 2)) + 1) * (tempDiff / 2) + previousSeasonModifier;
+		return MathUtil.round((float) seasonModifier, 2);
 	}
 }
